feat(models): add new models for video submissions, answers, questions, rounds, score logs, and contestant management
