+++ app/Models/User.php
@@ -1,22 +1,15 @@
 <?php
-// filepath: /c:/Laravel/OlympicTech_Backend_2025/app/Models/User.php
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
-use Laravel\Sanctum\HasApiTokens;
+
 
 class User extends Authenticatable
 {
-    use HasApiTokens, HasFactory, Notifiable;
+    use Notifiable;
 
-    /**
-     * Các trường có thể gán hàng loạt (Mass Assignment)
-     *
-     * @var array<int, string>
-     */
     protected $fillable = [
         'username',
         'email',
@@ -24,31 +17,13 @@ class User extends Authenticatable
         'role',
     ];
 
-    /**
-     * Các trường cần ẩn khi JSON hóa dữ liệu (Hidden Attributes)
-     *
-     * @var array<int, string>
-     */
     protected $hidden = [
         'password',
         'remember_token',
     ];
 
-    /**
-     * Các kiểu dữ liệu cần chuyển đổi (Casts)
-     *
-     * @var array<string, string>
-     */
-    protected $casts = [
-        'email_verified_at' => 'datetime',
-        'password' => 'hashed',
-    ];
-
-    /**
-     * Lấy các nhóm mà người dùng làm giám khảo
-     */
-    public function judgedGroups()
+    public function groups()
     {
-        return $this->hasMany(Groups::class, 'judge_id');
+        return $this->hasMany(Group::class, 'judge_id');
     }
 }
